rename vm to vm_logic
