+++ src/vm/mod.rs
@@ -1,7 +1,7 @@
 pub mod chunk;
 pub mod compiler;
-pub mod vm;
+pub mod vm_logic;
 
 pub use chunk::{Chunk, OpCode, VmValue};
 pub use compiler::{CompileError, Compiler};
-pub use vm::{Vm, VmError};
+pub use vm_logic::{Vm, VmError};
