+++ src/vm/vm_logic.rs
